requirements.txt
fastapi>=0.115.0
itsdangerous>=2.2.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
# pbkdf2_sha256 only — avoids passlib + bcrypt 4.1+ __about__ incompatibility
passlib>=1.7.4
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
requests>=2.31.0
